requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
Flask==2.0.3
Flask-RESTful==0.3.9
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.3
google-cloud-logging==3.0.0
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
isodate==0.6.1
itsdangerous==2.1.1
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.1.0
packaging==21.3
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
rsa==4.8
serde==0.8.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
urllib3==1.26.8
Werkzeug==2.0.3
